Job data enhancements
This change introduces the ability to add arbitrary properties
of any job data object to the table as a column. A JSON query
implementation is used to do this. Queries are relative to the
"root" of the json data object. These can be viewed in the API
browser of AWX if reference is needed.

Also added table connections.